Arreglo de pequeños errores

+++ app/models/cliente.py
@@ -5,6 +5,6 @@ class Cliente(db.Model):
     idcliente = db.Column(db.Integer, primary_key=True, autoincrement=True)
     nombrecte = db.Column(db.String(255), nullable=False)
     direccioncte = db.Column(db.String(255), nullable=False)
-    telefonocte = db.Column(db.Integer, nullable=False)
+    telefonocte = db.Column(db.BigInteger, nullable=False)
 
     facturas = db.relationship("Factura", back_populates="cliente")
